+++ ethers-contract/ethers-contract-abigen/src/rawabi.rs
@@ -8,6 +8,8 @@ use serde::{
 };
 
 /// Contract ABI as a list of items where each item can be a function, constructor or event
+#[derive(Debug, Clone, Serialize)]
+#[serde(transparent)]
 pub struct RawAbi(Vec<Item>);
 
 impl IntoIterator for RawAbi {
@@ -82,23 +84,31 @@ pub struct Item {
     pub name: Option<String>,
     #[serde(default)]
     pub outputs: Vec<Component>,
+    // required to satisfy solidity events
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub anonymous: Option<bool>,
 }
 
-/// Either
+/// Either an input/output or a nested component of an input/output
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Component {
     #[serde(rename = "internalType", default, skip_serializing_if = "Option::is_none")]
     pub internal_type: Option<String>,
+    #[serde(default)]
     pub name: String,
     #[serde(rename = "type")]
     pub type_field: String,
     #[serde(default)]
     pub components: Vec<Component>,
+    /// Indexed flag. for solidity events
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub indexed: Option<bool>,
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use ethers_core::abi::Abi;
 
     #[test]
     fn can_parse_raw_abi() {
@@ -111,5 +121,23 @@ mod tests {
         const VERIFIER_ABI: &str =
             include_str!("../../tests/solidity-contracts/verifier_abi_hardhat.json");
         let _ = serde_json::from_str::<RawAbi>(VERIFIER_ABI).unwrap();
+    }
+
+    /// due to ethabi's limitations some may be stripped when ethers-solc generates the abi, such as
+    /// the name of the component
+    #[test]
+    fn can_parse_ethers_solc_generated_abi() {
+        let s = r#"[{"type":"function","name":"greet","inputs":[{"internalType":"struct Greeter.Stuff","name":"stuff","type":"tuple","components":[{"type":"bool"}]}],"outputs":[{"internalType":"struct Greeter.Stuff","name":"","type":"tuple","components":[{"type":"bool"}]}],"stateMutability":"view"}]"#;
+        let _ = serde_json::from_str::<RawAbi>(s).unwrap();
+    }
+
+    #[test]
+    fn can_ethabi_round_trip() {
+        let s = r#"[{"anonymous":false,"inputs":[{"indexed":true,"internalType":"uint64","name":"number","type":"uint64"}],"name":"MyEvent","type":"event"},{"inputs":[],"name":"greet","outputs":[],"stateMutability":"nonpayable","type":"function"}]"#;
+
+        let raw = serde_json::from_str::<RawAbi>(s).unwrap();
+        let abi = serde_json::from_str::<Abi>(s).unwrap();
+        let de = serde_json::to_string(&raw).unwrap();
+        assert_eq!(abi, serde_json::from_str::<Abi>(&de).unwrap());
     }
 }