+++ ethers-etherscan/src/contract.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 
 use ethers_core::abi::{Abi, Address};
 
-use crate::{Client, Response, Result};
+use crate::{Client, EtherscanError, Response, Result};
 
 /// Arguments for verifying contracts
 #[derive(Debug, Clone, Serialize)]
@@ -236,6 +236,9 @@ impl Client {
     pub async fn contract_abi(&self, address: Address) -> Result<Abi> {
         let query = self.create_query("contract", "getabi", HashMap::from([("address", address)]));
         let resp: Response<String> = self.get_json(&query).await?;
+        if resp.result.starts_with("Contract source code not verified") {
+            return Err(EtherscanError::ContractCodeNotVerified(address))
+        }
         Ok(serde_json::from_str(&resp.result)?)
     }
 
@@ -268,7 +271,7 @@ mod tests {
     use serial_test::serial;
 
     use ethers_core::types::Chain;
-    use ethers_solc::{MinimalCombinedArtifacts, Project, ProjectPathsConfig};
+    use ethers_solc::{Project, ProjectPathsConfig};
 
     use crate::{contract::VerifyContract, tests::run_at_least_duration, Client};